Remove need for token based depoyment. (#335)
This pull request includes several important changes to the GitHub
Actions workflow for releases. The changes focus on configuring the
environment, updating permissions, and improving the build and publish
steps.

### Workflow configuration:

*
[`.github/workflows/release.yml`](diffhunk://#diff-87db21a973eed4fef5f32b267aa60fcee5cbdf03c67fafdc2a9b553bb0b15f34R13-R17):
Added the `environment: release` setting to the `deploy` job to specify
the environment for the release.

### Permissions:

*
[`.github/workflows/release.yml`](diffhunk://#diff-87db21a973eed4fef5f32b267aa60fcee5cbdf03c67fafdc2a9b553bb0b15f34R13-R17):
Added `id-token: write` permission, which is mandatory for trusted
publishing.

### Build and publish steps:

*
[`.github/workflows/release.yml`](diffhunk://#diff-87db21a973eed4fef5f32b267aa60fcee5cbdf03c67fafdc2a9b553bb0b15f34L35-R44):
Split the build and publish steps into two separate actions. The build
step now only builds the package distribution, and the publish step uses
the `pypa/gh-action-pypi-publish@release/v1` action to publish the
package distributions to PyPI.